Merge dev

+++ src/Engine/XrEngine.Compression/TextureCompressor.cs
@@ -1,5 +1,5 @@
 ﻿using System.Diagnostics;
-using System.Security.Cryptography;
+using XrEngine.Helpers;
 
 namespace XrEngine.Compression
 {
@@ -9,24 +9,24 @@ namespace XrEngine.Compression
         public int Align;
         public TextureCompressionFormat Format;
         public uint BlockSize;
+        public bool RequireRgba;
     }
 
     public class TextureCompressor
     {
-        readonly object _cacheLock = new object();
-        readonly Dictionary<string, Task<IList<TextureData>>> _encodeTasks = [];
-        readonly SemaphoreSlim _dictMutex = new(1, 1);
-        readonly SemaphoreSlim _encodeLimit = new(3, 3);
+        readonly object _cacheLock = new();
+        readonly AsyncTaskDispatcher _dispatcher = new AsyncTaskDispatcher(3, ThreadPriority.Lowest);
         bool _cacheCleared;
 
-        public static TextureCompressionInfo EncodeAstc(bool isNormalMap, float quality, uint blockSize)
+        public static TextureCompressionInfo EncodeAstc(bool isNormalMap, float quality, uint blockSize, int threadPriority)
         {
             return new TextureCompressionInfo
             {
                 Format = TextureCompressionFormat.Astc,
                 Align = 1,
                 BlockSize = blockSize,
-                Encode = data => AstcCompressor.Encode(data, isNormalMap, quality, blockSize)
+                RequireRgba = true,
+                Encode = data => AstcCompressor.Encode(data, isNormalMap, quality, blockSize, threadPriority)
             };
         }
 
@@ -41,73 +41,34 @@ namespace XrEngine.Compression
             };
         }
 
-
-        public async Task<IList<TextureData>> EncodeAsync(TextureData data, int mipsLevels, TextureCompressionInfo compressor)
+        public Task<IList<TextureData>> EncodeAsync(TextureData data, int mipsLevels, TextureCompressionInfo compressor, uint handle)
         {
-            var hash = TextureHash(data, compressor);
+            var hash = TextureHash(data, compressor, mipsLevels);
 
-            Task<IList<TextureData>> task;
-
-            await _dictMutex.WaitAsync().ConfigureAwait(false);
-            try
-            {
-                if (!_encodeTasks.TryGetValue(hash, out task!))
-                {
-                    task = Task.Run(async () =>
-                    {
-                        await _encodeLimit.WaitAsync().ConfigureAwait(false);
-                        try
-                        {
-                            return Encode(data, mipsLevels, hash, compressor);
-                        }
-                        finally
-                        {
-                            _encodeLimit.Release();
-                        }
-                    });
-
-                    _encodeTasks[hash] = task;
-                }
-            }
-            finally
-            {
-                _dictMutex.Release();
-            }
-
-            try
-            {
-                return await task.ConfigureAwait(false);
-            }
-            finally
-            {
-                await _dictMutex.WaitAsync().ConfigureAwait(false);
-                try
-                {
-                    _encodeTasks.Remove(hash);
-                }
-                finally
-                {
-                    _dictMutex.Release();
-                }
-            }
+            return _dispatcher.ExecuteAsync(() => Encode(data, mipsLevels, hash, compressor, handle), hash);
         }
 
-
-        public static string TextureHash(TextureData data, TextureCompressionInfo compressor)
+        public static string TextureHash(TextureData data, TextureCompressionInfo compressor, int mipsLevels)
         {
-            return Convert.ToHexString(MD5.HashData(data.Data!.AsSpan())) + "_" + compressor.Format + "_" + compressor.BlockSize + "_v6";
+            var dataHash = HashBuilder.Instance.Compute(data.Content!.AsSpan());
+
+            return $"{dataHash:X16}_{compressor.Format}_{compressor.BlockSize}_{mipsLevels}_v7";
         }
 
         public void ClearCache()
         {
             if (_cacheCleared)
                 return;
-            foreach (var file in Directory.GetFiles(CachePath!))
+
+            if (!Directory.Exists(CachePath))
+                return;
+
+            foreach (var file in Directory.GetFiles(CachePath))
                 File.Delete(file);
             _cacheCleared = true;
         }
 
-        public IList<TextureData> Encode(TextureData data, int mipsLevels, string? hash, TextureCompressionInfo compressor)
+        public IList<TextureData> Encode(TextureData data, int mipsLevels, string? hash, TextureCompressionInfo compressor, uint handle)
         {
             IList<TextureData>? result = null;
 
@@ -115,19 +76,33 @@ namespace XrEngine.Compression
 
             string? cacheFile = null;
 
-            hash ??= TextureHash(data, compressor);
+            string? validFile = null;
+
+            hash ??= TextureHash(data, compressor, mipsLevels);
 
             if (CachePath != null)
             {
+                //ClearCache();
+
                 cacheFile = Path.Combine(CachePath, hash + ".pvr");
+                validFile = cacheFile + ".valid";
 
                 lock (_cacheLock)
                 {
-                    if (File.Exists(cacheFile))
+                    if (File.Exists(cacheFile) && File.Exists(validFile))
                     {
-                        using var readStream = File.OpenRead(cacheFile);
-                        result = PvrTranscoder.Instance.LoadTexture(readStream);
-                        isCached = true;
+                        try
+                        {
+                            using var readStream = File.OpenRead(cacheFile);
+                            result = PvrTranscoder.Instance.LoadTexture(readStream);
+                            isCached = true;
+                            Log.Debug(this, "Loaded from cache: {0} - {1}", handle, Path.GetFileName(cacheFile));
+                        }
+                        catch (Exception ex)
+                        {
+                            Log.Warn(this, "Invalid compression cache '{0}':\n{1}", cacheFile, ex);
+                            File.Delete(validFile);
+                        }
                     }
                 }
             }
@@ -145,11 +120,29 @@ namespace XrEngine.Compression
                     var width = (int)MathF.Max(1, data.Width >> level);
                     var height = (int)MathF.Max(1, data.Height >> level);
 
-                    var resizeData = ImageUtils.ResizeV2(lastData, width, height);
+                    var resizeData = ImageUtils.Resize(lastData, width, height);
 
                     Log.Info(this, "Compressing mip {0} mipsLevels width {1} height {2}", level, width, height);
 
-                    var packData = ImageUtils.Pack(resizeData, compressor.Align);
+                    TextureData packData;
+
+                    if (compressor.RequireRgba)
+                    {
+                        if (resizeData.Format == TextureFormat.Rgba16)
+                        {
+                            Debug.Assert(compressor.Align == 1);
+                            packData = ImageUtils.ConvertRgba16ToRgba32F(resizeData);
+                        }
+                        else if (resizeData.Format == TextureFormat.RgbFloat32)
+                        {
+                            Debug.Assert(compressor.Align == 1);
+                            packData = ImageUtils.ConvertRgb32FToRgba16F(resizeData);
+                        }
+                        else
+                            packData = ImageUtils.PackToRgba8(resizeData, compressor.Align);
+                    }
+                    else
+                        packData = ImageUtils.Pack(resizeData, compressor.Align);
 
                     var newData = compressor.Encode(packData);
 
@@ -172,8 +165,20 @@ namespace XrEngine.Compression
                     lock (_cacheLock)
                     {
                         Directory.CreateDirectory(Path.GetDirectoryName(cacheFile)!);
+
+                        if (File.Exists(cacheFile))
+                            File.Delete(cacheFile);
+
+                        if (File.Exists(validFile))
+                            File.Delete(validFile);
+
                         using var writeStream = File.OpenWrite(cacheFile);
+
                         PvrTranscoder.Instance.SaveTexture(writeStream, result);
+
+                        writeStream.Close();
+
+                        File.WriteAllText(validFile!, "");
                     }
                 }
             }
@@ -184,16 +189,14 @@ namespace XrEngine.Compression
             {
                 if (mipsLevels == 0)
                     item.MipLevel = data.MipLevel;
-                item.Face = data.Face;
+                item.Layer = data.Layer;
                 item.Depth = data.Depth;
             }
 
             return result;
         }
 
-
         public string? CachePath { get; set; }
-
 
         public static readonly TextureCompressor Instance = new();
